+++ lib/graphql_operations/queries/get_all_products_from_collection_by_id.dart
@@ -17,6 +17,11 @@ query($id : ID!, $cursor : String, $sortKey: ProductCollectionSortKeys){
         edges {
           cursor
           node {
+          options(first: 50) {
+            id
+            name
+            values
+            } 
             availableForSale
             collections(first: 250) {
               edges {