+++ lib/graphql_operations/queries/get_checkout_information.dart
@@ -15,15 +15,6 @@ query($id: ID!){
         }
         id
       }
-      availableShippingRates {
-        ready
-        shippingRates {
-          handle
-          title
-          priceV2 {
-            amount
-            currencyCode
-       }}}
       requiresShipping
       completedAt
       createdAt