+++ lib/graphql_operations/queries/get_products.dart
@@ -7,6 +7,11 @@ query($cursor : String){
     edges {
       cursor
       node {
+      options(first: 50) {
+            id
+            name
+            values
+            } 
         variants(first: 250) {
           edges {
             node {