+++ lib/graphql_operations/queries/get_products_by_ids.dart
@@ -2,6 +2,11 @@ const String getProductsByIdsQuery = r'''
 query($ids : [ID!]!){
   nodes(ids: $ids) {
     ... on Product {
+    options(first: 50) {
+            id
+            name
+            values
+            } 
       id
       handle
       collections(first: 250) {