+++ lib/graphql_operations/queries/get_x_products_on_query_after_cursor.dart
@@ -3,6 +3,11 @@ query( $cursor: String, $limit : Int, $sortKey : ProductSortKeys, $query: String
   products(query: $query, first: $limit, after: $cursor, sortKey: $sortKey, reverse: $reverse) {
     edges {
       node {
+      options(first: 50) {
+            id
+            name
+            values
+            } 
         id
         handle
         availableForSale